# TaxDeedFlow/public/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tax Deed Flow API",
    "version": "1.0.0",
    "description": "Public REST API for accessing tax deed property data, county information, and risk analysis. Requires API key authentication.",
    "contact": {
      "name": "Tax Deed Flow Support",
      "url": "https://taxdeedflow.com/developer"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://taxdeedflow.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/counties": {
      "get": {
        "tags": ["Counties"],
        "summary": "List counties",
        "description": "Retrieve a paginated list of counties with auction information",
        "operationId": "listCounties",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "state_code",
            "in": "query",
            "description": "Filter by state code (e.g., PA, FL, TX)",
            "schema": {
              "type": "string",
              "pattern": "^[A-Z]{2}$"
            }
          },
          {
            "name": "has_upcoming_sale",
            "in": "query",
            "description": "Filter counties with upcoming auctions",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of results to return",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of results to skip",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCountiesResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitExceeded"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/developer/keys": {
      "post": {
        "tags": ["API Keys"],
        "summary": "Create API key",
        "description": "Generate a new API key with specified permissions and rate limit tier. The plaintext API key is only shown once.",
        "operationId": "createApiKey",
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApiKeyResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "get": {
        "tags": ["API Keys"],
        "summary": "List API keys",
        "description": "Retrieve all API keys for the authenticated user",
        "operationId": "listApiKeys",
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListApiKeysResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/developer/keys/{id}": {
      "delete": {
        "tags": ["API Keys"],
        "summary": "Revoke API key",
        "description": "Revoke an API key, preventing future use. This action cannot be undone.",
        "operationId": "revokeApiKey",
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "API key ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key revoked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeApiKeyResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "API key already revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for accessing public v1 endpoints. Generate keys at /developer/keys"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session-based authentication for developer management endpoints"
      }
    },
    "schemas": {
...
      "County": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "County unique identifier"
          },
          "county_name": {
            "type": "string",
            "description": "County name"
          },
          "state_code": {
            "type": "string",
            "pattern": "^[A-Z]{2}$",
            "description": "Two-letter state code"
          },
          "state_name": {
            "type": "string",
            "description": "Full state name"
          },
          "research_status": {
            "type": "string",
            "nullable": true,
            "description": "Research completion status"
          },
          "last_researched_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Last research timestamp"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Record creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Record last updated timestamp"
          }
        }
      },
      "CountyWithDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/County"
          },
          {
            "type": "object",
            "properties": {
              "property_count": {
                "type": "integer",
                "description": "Number of properties in county"
              },
              "upcoming_sale": {
                "type": "object",
                "nullable": true,
                "properties": {
                  "sale_date": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Auction date"
                  },
                  "sale_type": {
                    "type": "string",
                    "description": "Type of sale (e.g., tax deed, tax lien)"
                  },
                  "registration_deadline": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true,
                    "description": "Registration deadline"
                  }
                }
              }
            }
          }
        ]
      },
      "GetCountiesResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CountyWithDetails"
            }
          },
          "count": {
            "type": "integer",
            "description": "Total number of matching counties"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
...
      "CreateApiKeyRequest": {
        "type": "object",
        "required": ["name", "permissions"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Human-readable name for the API key"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["read", "write", "delete", "admin"]
            },
            "minItems": 1,
            "description": "Array of permissions to grant"
          },
          "rate_limit_tier": {
            "type": "string",
            "enum": ["free", "pro", "enterprise", "unlimited"],
            "default": "free",
            "description": "Rate limit tier for this API key"
          }
        }
      },
      "CreateApiKeyResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ApiKeyWithSecret"
          },
          "message": {
            "type": "string",
            "description": "Success message with warning to save the key"
          }
        }
      },
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "API key name"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Granted permissions"
          },
          "rate_limit_tier": {
            "type": "string",
            "enum": ["free", "pro", "enterprise", "unlimited"],
            "description": "Rate limit tier"
          },
          "last_used_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Last usage timestamp"
          },
          "request_count": {
            "type": "integer",
            "description": "Total request count"
          },
          "revoked_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Revocation timestamp"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiKeyWithSecret": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiKey"
          },
          {
            "type": "object",
            "properties": {
              "api_key": {
                "type": "string",
                "description": "Plaintext API key - only shown once on creation"
              }
            }
          }
        ]
      },
      "ListApiKeysResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiKey"
            }
          },
          "count": {
            "type": "integer",
            "description": "Total number of API keys"
          }
        }
      },
      "RevokeApiKeyResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer",
            "description": "Number of results per page"
          },
          "offset": {
            "type": "integer",
            "description": "Number of results skipped"
          },
          "has_more": {
            "type": "boolean",
            "description": "Whether more results are available"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request - invalid parameters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Bad Request",
              "message": "Invalid parameter: limit cannot exceed 1000"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - missing or invalid API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Unauthorized",
              "message": "Missing or invalid API key"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Forbidden",
              "message": "Your API key does not have read permission"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found - resource does not exist",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Not Found",
              "message": "Property not found"
            }
          }
        }
      },
      "RateLimitExceeded": {
        "description": "Rate limit exceeded",
        "headers": {
          "X-RateLimit-Limit": {
            "description": "The rate limit ceiling for this endpoint",
            "schema": {
              "type": "integer"
            }
          },
          "X-RateLimit-Remaining": {
            "description": "The number of requests remaining in the current window",
            "schema": {
              "type": "integer"
            }
          },
          "X-RateLimit-Reset": {
            "description": "Unix timestamp when the rate limit resets",
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Rate Limit Exceeded",
              "message": "Rate limit exceeded. Limit: 1000/hour. Resets at 2024-01-23T15:00:00Z"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Server Error",
              "message": "An unexpected error occurred"
            }
          }
        }
      }
    }
  }
}